Add Ubuntu 24.04
This uses no-cloud-init; and is a very interesting way to quickly get
an image going.

Starting point is a live-image, combined with a cloud-init package
that can be fed installation requirements, ssh keys, user
definition. This is exactly what we were doing via serial, but in a
much less fragile way.